Refactor requests checks
