+++ packages/sanes-chrome-extension/src/routes/share-identity/index.tsx
@@ -1,22 +1,12 @@
-import { Location } from 'history';
 import { ToastContext } from 'medulas-react-components/lib/context/ToastProvider';
-import { ToastVariant } from 'medulas-react-components/lib/context/ToastProvider/Toast';
 import * as React from 'react';
 import { RouteComponentProps, withRouter } from 'react-router';
 import { RequestContext } from '../../context/RequestProvider';
 import { history } from '../../store/reducers';
 import { REQUEST_ROUTE } from '../paths';
-import { REQUEST_FIELD } from '../requests/components/RequestList';
+import { checkRequest } from '../requests';
 import RejectRequest from './components/RejectRequest';
 import ShowRequest from './components/ShowRequest';
-
-function getIdFrom(location: Location): number | undefined {
-  if (!location || !location.state) {
-    return undefined;
-  }
-
-  return location.state[REQUEST_FIELD];
-}
 
 const ShareIdentity = ({ location }: RouteComponentProps): JSX.Element => {
   const [action, setAction] = React.useState<'show' | 'reject'>('show');
@@ -24,12 +14,7 @@ const ShareIdentity = ({ location }: RouteComponentProps): JSX.Element => {
   const requestContext = React.useContext(RequestContext);
 
   const request = requestContext.firstRequest;
-  const expectedId = getIdFrom(location);
-
-  if (!request || typeof expectedId === undefined || expectedId !== request.id) {
-    toast.show('Error: Request not identified', ToastVariant.ERROR);
-    history.push(REQUEST_ROUTE);
-  }
+  checkRequest(request, location, toast);
 
   const showRequestView = (): void => setAction('show');
   const showRejectView = (): void => setAction('reject');
